Move `QuizDelegate` to it's own file.
