+++ QuizCore/QuizCore/QuizDelegate.swift
@@ -1,0 +1,18 @@
+//
+//  QuizDelegate.swift
+//  QuizCore
+//
+//  Created by Saba Khutsishvili on 20.09.21.
+//
+
+import Foundation
+
+public protocol QuizDelegate {
+    
+    associatedtype Question: Hashable
+    associatedtype Answer
+    
+    func handle(question: Question, answerCallback: @escaping (Answer) -> Void)
+    func handle(result: QuizResult<Question, Answer>)
+    
+}
