+++ QuizCore/QuizCore/Router.swift
@@ -7,16 +7,6 @@
 
 import Foundation
 
-public protocol QuizDelegate {
-    
-    associatedtype Question: Hashable
-    associatedtype Answer
-    
-    func handle(question: Question, answerCallback: @escaping (Answer) -> Void)
-    func handle(result: QuizResult<Question, Answer>)
-    
-}
-
 @available(*, deprecated)
 public protocol Router {
